media-tools/recorder/transcripts/requirements.txt
# 🎤 SPEECH RECORDER TRANSCRIPTION TOOL
# Python dependencies for Mac/Windows/Linux
# Usage: pip install -r requirements.txt

openai-whisper==20231117
moviepy==1.0.3

# Optional: For faster processing (if you have GPU)
# torch==2.1.0
# torchaudio==2.1.0